refactor: create api service adapter out of institution service
To use different APIs, the institution-service will not know API
specific information. Instead, there is an interface at the
api-service with 2 calls. This is a preparation for a second
api-service in a derivative kulturfinder.
